AAAAA OPENCL WORKS HOLY FUCK

+++ includes/fract_structs.h
@@ -1,35 +1,35 @@
 #ifndef _FRACT_STRUCTS_H
 # define _FRACT_STRUCTS_H
 
-typedef struct	s_mouse
+typedef struct		s_mouse
 {
-	int			press;
-	double		x;
-	double		y;
-	double 		Re;
-	double 		Im;
-}				t_mouse;
+	int				press;
+	float			x;
+	float			y;
+	float 			Re;
+	float 			Im;
+}					t_mouse;
 
 typedef struct 		s_position
 {
-	double  		shift_x;
-	double  		shift_y;
-	double 			zoom;
+	float  			shift_x;
+	float  			shift_y;
+	float 			zoom;
 }					t_pos;
 
 typedef struct 		s_mandelbrot
 {
-	double 			c_re;
-	double  		c_im;
-	double 			x;
-	double 			y;
-	double 			x_sqr;
-	double 			y_sqr;
-	double 			re_min;
-	double 			re_max;
-	double 			im_min;
-	double 			im_max;
-	double 			im_max_start;
+	float 			c_re;
+	float  			c_im;
+	float 			x;
+	float 			y;
+	float 			x_sqr;
+	float 			y_sqr;
+	float 			re_min;
+	float 			re_max;
+	float 			im_min;
+	float 			im_max;
+	float 			im_max_start;
 	int 			max_iter;
 	t_pos			*pos;
 }					t_mandel;
