+++ srcs/color_managment.c
@@ -10,7 +10,7 @@ int				color(int iter, int max_iter)
 	int			r;
 	int			g;
 	int			b;
-	double 		t;
+	float 		t;
 
 
 	r = gradient((START_COLOR >> 16) & 0xFF, (END_COLOR >> 16) & 0xFF, iter);
@@ -18,7 +18,7 @@ int				color(int iter, int max_iter)
 	b = gradient(START_COLOR & 0xFF, END_COLOR & 0xFF, iter);
 
 	t = max_iter;
-//	t = (double)iter / (double)max_iter;
+//	t = (float)iter / (float)max_iter;
 //	r = (int)(9 * (1 - t) * pow(t, 3) * 255);
 //	g = (int)(15 * pow((1 - t), 2) * pow(t, 2) * 255);
 //	b = (int)(8.5 * pow((1 - t), 3) * t * 255);
