+++ srcs/init.c
@@ -18,9 +18,9 @@ static	t_pos		*init_position(void)
 	t_pos	*pos;
 
 	pos = (t_pos *)malloc(sizeof(t_pos));
-	pos->shift_x = -0.496;
-	pos->shift_y = 0.496;
-	pos->zoom = 1.0;
+	pos->shift_x = (float)-0.496;
+	pos->shift_y = (float)0.496;
+	pos->zoom = 1;
 	return (pos);
 }
 
@@ -34,9 +34,9 @@ static t_mandel		*init_mandelbrot(void)
 	man->c_re = 0;
 	man->x = 0;
 	man->y = 0;
-	man->re_max = 2.0;
-	man->re_min = -2.0;
-	man->im_min = -2.0;
+	man->re_max = 2;
+	man->re_min = -2;
+	man->im_min = -2;
 	man->im_max = man->im_min + (man->re_max - man->re_min) * HEIGHT / WIDTH;
 	man->im_max_start = man->im_max;
 	man->max_iter = K_START;
